requirements.txt
discord.py>=2.3.0
open-interpreter>=0.2.0
pyautogui>=0.9.54
psutil>=5.9.0
selenium>=4.15.0
openai-whisper>=20231117
python-dotenv>=1.0.0
pyttsx3>=2.90
Pillow>=10.0.0